requirements.txt
dj-database-url==0.4.2
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.3
whitenoise==3.3.0
drfdocs==0.0.11